move limit input within widget
