+++ src/components/widgets/RecentTimes.tsx
@@ -17,17 +17,20 @@ import { secondsToDate } from "../../utils/DateUtils";
 import RadioButtons from "./RadioButtons";
 
 interface RecentTimesProps {
-  limit: number;
+  defaultLimit?: number;
+  canChangeLimit?: boolean;
 }
 
-const RecentTimes = ({ limit }: RecentTimesProps) => {
+const RecentTimes = ({ defaultLimit, canChangeLimit = false }: RecentTimesProps) => {
   const { lang } = useContext(I18nContext);
   const [records, setRecords] = useState(false);
   const metadata = useContext(MetadataContext);
 
+  const [limit, setLimit] = useState(defaultLimit ?? 30);
+
   const { isLoading: recentTimesLoading, data } = useApi(
     () =>
-      Score.getRecent(limit, records).then((scores) =>
+      Score.getRecent(limit ?? 30, records).then((scores) =>
         scores.reduce(
           (acc, data, index) => {
             const track = metadata.getTrackById(data.trackId);
@@ -102,15 +105,40 @@ const RecentTimes = ({ limit }: RecentTimesProps) => {
           : translate("recentTimesRecentTimesHeading", lang)
       }
     >
-      <RadioButtons
-        data={[
-          { text: "All", value: false },
-          { text: "Records", value: true },
-        ]}
-        state={records}
-        setState={setRecords}
-        className="recent-times-input"
-      />
+      <div style={{ padding: "5px" }}>
+        {canChangeLimit && (
+          <div
+            className="module-row"
+            style={{ gap: "5px", alignItems: "baseline", paddingBottom: "5px" }}
+          >
+            <span>{translate("trackListPageMoreTimes", lang)}</span>
+            <input
+              type="number"
+              style={{ flexGrow: 1 }}
+              defaultValue={30}
+              onChange={(e) => {
+                let value = parseInt(e.target.value);
+                if (isNaN(value) || value < 31) {
+                  e.target.value = String((value = 30));
+                }
+                setLimit(value);
+              }}
+              step="30"
+              min="30"
+              pattern="[0-9]*"
+            />
+          </div>
+        )}
+        <RadioButtons
+          data={[
+            { text: translate("constantCountryRankingsTopEnumAll", lang), value: false },
+            { text: translate("constantCountryRankingsTopEnumRecords", lang), value: true },
+          ]}
+          state={records}
+          setState={setRecords}
+          className="recent-times-input"
+        />
+      </div>
       <Deferred isWaiting={recentTimesLoading || metadata.isLoading}>
         <ArrayTable
           className="recent-times-table"
